chatbot/requirements.txt
streamlit==1.39.0
streamlit-cognito-auth==1.3.1
boto3==1.34.146
markdown==3.6
get-code-from-markdown==1.0.0
defusedxml==0.7.1
requests==2.32.3
pypdf==5.1.0
langchain==0.3.7
langchain-community==0.3.3
unstructured==0.17.2
python-pptx==1.0.2
pyshorteners==1.0.1
python-dotenv==1.0.1
google-cloud-storage==2.14.0
google-cloud-firestore==2.15.0
google-cloud-aiplatform==1.50.0
# A2A dependencies
starlette==0.35.1
uvicorn==0.28.0
httpx==0.28.1
pydantic==2.7.4
pyjwt==2.10.1
cryptography==44.0.3
# Gemini dependencies
google-generativeai==0.8.5
